backend/pyproject.toml
[project]
name = "digidocs-mgt-backend"
version = "0.1.0"
description = "Backend service for digidocs management platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "celery>=5.5.0",
  "fastapi>=0.116.0",
  "pydantic-settings>=2.10.0",
  "psycopg[binary]>=3.2.0",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.20",
  "redis>=6.0.0",
  "sqlalchemy>=2.0.40",
  "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.4.0",
  "pytest-asyncio>=1.1.0",
  "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

